extract evaluator into another module
