+++ xiao3sh/src/main.rs
@@ -1,57 +1,14 @@
-use std::collections::HashMap;
-use std::io::{self, Error, ErrorKind, Write};
-use std::process::Command;
-
-mod builtins;
-
-fn split_line(line: &str) -> impl Iterator<Item = &str> {
-    line.trim().split_whitespace()
-}
-
-fn launch(cmd: &str, args: &[String]) -> io::Result<()> {
-    match Command::new(cmd).args(args).status() {
-        Ok(_) => Ok(()),
-        Err(err) => {
-            println!("{}", err);
-            Ok(())
-        }
-    }
-}
-
-fn execute(
-    args: &[String],
-    builtins: &HashMap<String, fn(&[String]) -> io::Result<()>>,
-) -> io::Result<()> {
-    if args.len() < 1 {
-        return Err(Error::new(ErrorKind::InvalidInput, "command is not given"));
-    }
-    let cmd = &args[0];
-    match builtins.get(cmd) {
-        Some(func) => func(&args[1..]),
-        None => launch(&cmd, &args[1..]),
-    }
-}
-
-fn conv<'a, 'b, I>(x: I) -> Vec<String>
-where
-    I: Iterator<Item = &'a str>,
-{
-    let mut r: Vec<String> = Vec::new();
-    for e in x {
-        let s = e.to_string().clone();
-        r.push(s);
-    }
-    r
-}
+use std::io::{self, Write};
+mod eval;
 
 fn main_loop() -> io::Result<()> {
     let mut input = String::new();
-    let builtins = builtins::create_builtins();
+    let eval = &eval::Evaluator::new();
     loop {
         print!("> ");
         io::stdout().flush().unwrap();
         io::stdin().read_line(&mut input)?;
-        execute(&conv(split_line(&input)), &builtins)?;
+        eval.execute(&input)?;
         input.truncate(0);
         io::stdout().flush().unwrap();
     }
